ensure that SendAllGossipDown is sent before forwarded broadcasts
Once LocalPeer.addConnection has been invoked, the connection is
accessible to gossip routing / transmission. GossipSenders are created
the first time this happens for any particular gossip channel. These
GossipSenders are initially blocked, waiting for a signal. As a result
they will just accumulate gossip via GossipData.Merge. This includes
the gossip emitted by SendAllGossipDown, which is invoked as part of
LocalPeer.AddConnection, just after addConnection.

We send the signal just after LocalPeer.AddConnection returns. So by
this time we can be sure that all the SendAllGossipDown gossip is
sitting the GossipSenders. And because GossipSenders always send such
"proper" gossip before broadcasts, it gets sent first on the
connection.

This is important for topology gossip because forwarded broadcasts
might be about peers which are only reachable by the remote if it has
an up-to-date(ish) topology.